refactor(gui): extract App CRUD + discovery worker into AppCrudMixin
Move app profile add/edit/delete handlers and the Refresh-Apps QThread
worker orchestration off WinpodxWindow into a private mixin so
main_window.py can focus on overall window orchestration:

- AppCrudMixin (src/winpodx/gui/_main_window_apps.py)
  _on_add_app, _on_edit_app, _on_delete_app, _reload_apps,
  _on_refresh_apps (the QThread spinup), _set_refresh_state,
  _on_refresh_succeeded, _cleanup_refresh_worker, _on_refresh_failed,
  _show_refresh_failure_dialog (multi-branch failure dialog).

WinpodxWindow now reads
`class WinpodxWindow(AppCrudMixin, InfoPageMixin, LogsMixin, MaintenanceMixin, QMainWindow)`.

The v0.1.9.1 QTimer.singleShot deferral for QMessageBox creation and
the v0.2.0.11 worker-cleanup ordering both come along as-is — they
were behaviour bug fixes encoded in the methods. No behavior change;
tests: 1096 passed.

main_window.py: 2016 -> 1828 lines (-188). Cumulative: 2745 -> 1828 (-33%).
